Make it possible to add from template to any open invoice for biller and client.

+++ add_rows.php
@@ -55,8 +55,8 @@ if (!sesWriteAccess()) {
     return;
 }
 
-$invoiceId = getPostOrQuery('id');
-$templateId = getPostOrQuery('template_id');
+$invoiceId = (int)getPostOrQuery('id');
+$templateId = (int)getPostOrQuery('template_id');
 
 if (!$invoiceId || !$templateId) {
     echo htmlPageStart();
